requirements.txt
pyaes==1.6.1
pyasn1==0.5.0
python-dateutil==2.8.2
rsa==4.9
six==1.16.0
Telethon==1.28.5
